refactor(scraper): package layout under scrape/ with main entrypoint
Split the monolithic scraper into scrape.core, scrape.sources, and
scrape.valuation. Keep main.py as the single CLI entry (uv run main.py).
Behavior and CI workflow are unchanged.

Made-with: Cursor
